fix: address Copilot PR #703 round-3 review comments
- Extract shared TimeWindow type, TIME_WINDOW_OPTIONS, and getWindowFrom
  into lib/timeWindows.ts; remove duplicate definitions from ScienceDataSection
  and DepthSection so both components stay in sync automatically

- usePersistentState: fix SSR hydration mismatch — initialize with initialValue
  (matches server render), then hydrate from sessionStorage in useLayoutEffect
  (client-only, runs before paint so there is no visible flash)

- ScienceDataSection: remove duplicate bottom fade gradient div; AccordionCells
  already renders its own fade overlay
