+++ apps/lrauv-dash2/lib/timeWindows.ts
@@ -1,0 +1,37 @@
+import { DateTime } from 'luxon'
+
+export type TimeWindow = 'latest' | '3d' | '7d' | 'deployment'
+
+export const TIME_WINDOW_OPTIONS: { id: TimeWindow; name: string }[] = [
+  { id: 'latest', name: 'Latest Dive' },
+  { id: '3d', name: '3 Days' },
+  { id: '7d', name: '7 Days' },
+  { id: 'deployment', name: 'Full Deployment' },
+]
+
+/**
+ * Returns the start timestamp (ms) for the given time window.
+ *
+ * For ended deployments `deploymentTo` is in the past, so relative windows
+ * (3d/7d) are anchored from the deployment end, not now. The anchor is
+ * clamped to now so future-padded end times on active deployments don't
+ * shift the window forward and silently drop older data.
+ */
+export const getWindowFrom = (
+  window: TimeWindow,
+  deploymentFrom: number,
+  deploymentTo?: number
+): number => {
+  const now = DateTime.utc().toMillis()
+  const anchor = Math.min(deploymentTo ?? now, now)
+  switch (window) {
+    case '3d':
+      return Math.max(deploymentFrom, anchor - 3 * 24 * 60 * 60 * 1000)
+    case '7d':
+      return Math.max(deploymentFrom, anchor - 7 * 24 * 60 * 60 * 1000)
+    case 'deployment':
+      return deploymentFrom
+    default:
+      return deploymentFrom
+  }
+}
